server/requirements.txt
pycryptodome>=3.18.0
SimpleWebSocketServer>=0.1.2
